Rename deploy script
