+++ tasks/index.ts
@@ -1,4 +1,4 @@
-import "./deployTevmos";
+import "./deploy";
 import "./upgrade";
 import "./status";
 import "./delegate";
